rework die system

+++ Assets/Scripts/Logic/Health/IHealth.cs
@@ -1,9 +1,7 @@
-using System;
-
 namespace Logic.Health
 {
 	public interface IHealth 
 	{
-		void TakeDamage(float damage);
+		void TakeDamage(int damage);
 	}
 }
