+++ Assets/Scripts/Logic/Health/IHealthAddable.cs
@@ -2,6 +2,6 @@ namespace Logic.Health
 {
 	public interface IHealthAddable : IHealth
 	{
-		void AddHealth(float value);
+		void AddHealth(int value);
 	}
 }
